@@ -1,44 +1,59 @@
 {
   "name": "bith-annotator",
-  "version": "0.1.0",
+  "version": "0.9.0",
+  "license": "AGPL-3.0",
+  "author": "Johannes Kepper",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/DomesticBeethoven/bith-annotator.git"
+  },
+  "homepage": "https://github.com/DomesticBeethoven/bith-annotator#readme",
+  "bugs": {
+    "url": "https://github.com/DomesticBeethoven/bith-annotator/issues"
+  },
   "scripts": {
     "serve": "vue-cli-service serve",
+    "build": "vue-cli-service build",
     "test": "standard --fix | snazzy",
-    "build": "vue-cli-service build"
+    "lint": "vue-cli-service lint"
   },
   "dependencies": {
-    "@inrupt/solid-client": "^1.19.0",
-    "@inrupt/solid-client-authn-browser": "^1.11.5",
-    "@recogito/annotorious-openseadragon": "^2.7.2",
+    "@inrupt/solid-client": "^1.23.2",
+    "@inrupt/solid-client-authn-browser": "^1.12.2",
+    "@recogito/annotorious-openseadragon": "^2.7.7",
     "@recogito/annotorious-toolbar": "^1.1.0",
-    "core-js": "^3.6.5",
-    "jsonld": "^5.2.0",
-    "meld-clients-core": "github:DomesticBeethoven/meld-clients-core",
-    "openseadragon": "^3.0.0",
-    "rdflib": "^2.2.18",
-    "redux": "^4.1.2",
+    "core-js": "^3.8.3",
+    "crypto-browserify": "^3.12.0",
+    "openseadragon": "^3.1.0",
+    "rdflib": "^2.2.19",
+    "redux": "^4.2.0",
     "redux-thunk": "^2.4.1",
-    "solid-auth-client": "^2.5.6",
-    "splitpanes": "^2.4.1",
-    "svg-drag-select": "^0.4.2",
-    "uuid": "^8.3.2",
-    "verovio": "^3.9.0",
-    "vue": "^2.6.11",
-    "vuex": "^3.4.0"
+    "splitpanes": "^3.1.1",
+    "verovio": "^3.11.0",
+    "vue": "^3.2.13",
+    "vuex": "^4.0.0"
   },
   "devDependencies": {
-    "@vue/cli-plugin-babel": "~4.5.15",
-    "@vue/cli-plugin-vuex": "~4.5.15",
-    "@vue/cli-service": "~4.5.15",
-    "node-sass": "^4.12.0",
-    "patch-package": "^6.4.7",
-    "sass-loader": "^8.0.2",
+    "@babel/core": "^7.12.16",
+    "@babel/eslint-parser": "^7.12.16",
+    "@vue/cli-plugin-babel": "~5.0.0",
+    "@vue/cli-plugin-eslint": "~5.0.0",
+    "@vue/cli-plugin-vuex": "~5.0.0",
+    "@vue/cli-service": "~5.0.0",
+    "@vue/eslint-config-standard": "^6.1.0",
+    "eslint": "^7.32.0",
+    "eslint-plugin-import": "^2.25.3",
+    "eslint-plugin-node": "^11.1.0",
+    "eslint-plugin-promise": "^5.1.0",
+    "eslint-plugin-vue": "^8.0.3",
+    "lint-staged": "^11.1.2",
+    "meld-clients-core": "github:DomesticBeethoven/meld-clients-core",
+    "sass": "^1.32.7",
+    "sass-loader": "^12.0.0",
     "snazzy": "^9.0.0",
-    "standard": "^16.0.4",
-    "vue-template-compiler": "^2.6.11"
+    "standard": "^17.0.0"
   },
-  "private": true,
-  "eslintIgnore": [
-    "**/public/resources/**/*.js"
-  ]
+  "gitHooks": {
+    "pre-commit": "lint-staged"
+  }
 }
